Simplify room migrations
